Treat an empty SITE_KEY as error
Recently I had a configuration error where the SITE_KEY remained unset
and the empty defaults made the widget render without the captcha and
subsequently fail every submission. While the form showed a validation
error, it took days until someone reported it.  Lack of logging made
tracking down this issue needlessly long as well.  When this widget is
required, having an empty SITE_KEY and/or SITE_SECRET is simply an error
and should be handled as such.

While here, simplify the widget and add the SITE_KEY as the hidden input
field's value= instead of creating a free floating <script> tag with a
single variable...

Remove the empty defaults, they dont make sense.
This diff is against 0.16.1